// frontend/src/api/schema.d.ts
export interface paths {
  "/api/v1/tags/{id}": {
    put: operations["update"];
    delete: operations["delete_1"];
  };
  "/api/v1/outfits/{id}": {
    get: operations["get"];
    put: operations["update_1"];
    delete: operations["delete_2"];
  };
  "/api/v1/outfits/{id}/items/reorder": {
    put: operations["reorder"];
  };
  "/api/v1/clothing/{id}": {
    get: operations["get_1"];
    put: operations["update_2"];
    delete: operations["delete_3"];
  };
  "/api/v1/categories/{id}": {
    put: operations["update_3"];
    delete: operations["delete_4"];
  };
  "/api/v1/calendar/{id}": {
    get: operations["get_2"];
    put: operations["update_4"];
    delete: operations["delete_5"];
  };
  "/api/v1/wear-logs": {
    post: operations["create"];
  };
  "/api/v1/tags": {
    get: operations["list"];
    post: operations["create_1"];
  };
  "/api/v1/outfits": {
    get: operations["list_1"];
    post: operations["create_2"];
  };
  "/api/v1/outfits/{id}/items": {
    post: operations["addItem"];
  };
  "/api/v1/clothing": {
    get: operations["list_2"];
    post: operations["create_3"];
  };
  "/api/v1/clothing/{id}/images": {
    post: operations["uploadImage"];
  };
  "/api/v1/categories": {
    get: operations["list_3"];
    post: operations["create_4"];
  };
  "/api/v1/calendar": {
    get: operations["range"];
    post: operations["create_5"];
  };
  "/api/v1/stats/overview": {
    get: operations["overview"];
  };
  "/api/v1/stats/most-worn": {
    get: operations["mostWorn"];
  };
  "/api/v1/stats/least-worn": {
    get: operations["leastWorn"];
  };
  "/api/v1/stats/clothing/{id}": {
    get: operations["forClothing"];
  };
  "/api/v1/images/**": {
    get: operations["proxy"];
  };
  "/api/v1/wear-logs/{id}": {
    delete: operations["delete"];
  };
  "/api/v1/outfits/{id}/items/{clothingId}": {
    delete: operations["removeItem"];
  };
  "/api/v1/clothing/{id}/images/{imageId}": {
    delete: operations["deleteImage"];
  };
}